requirements.txt
tensorflow==2.10.0
streamlit==1.29.0
opencv-python-headless==4.6.0.66
numpy
pytesseract
pillow
spacy==3.0.0
